requirements.txt
gradio
faiss-cpu
numpy
pandas
langchain
sentence-transformers
huggingface_hub
